+++ src/Objects/Vector.java
@@ -1,72 +1,72 @@
 package Objects;
 
 public class Vector {
-	private double x;
-	private double y;
-	private double z;
+	private float x;
+	private float y;
+	private float z;
 	
-	public Vector(double x, double y, double z) {
+	public Vector(float x, float y, float z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
 		normalize();
 	}
-	public double getX() {
+	public float getX() {
 		return x;
 	}
-	public void setX(double x) {
+	public void setX(float x) {
 		this.x = x;
 	}
-	public double getY() {
+	public float getY() {
 		return y;
 	}
-	public void setY(double y) {
+	public void setY(float y) {
 		this.y = y;
 	}
-	public double getZ() {
+	public float getZ() {
 		return z;
 	}
-	public void setZ(double z) {
+	public void setZ(float z) {
 		this.z = z;
 	}
 	
-	public double length(){
-		return Math.sqrt(Math.pow(getX(), 2) + Math.pow(getY(), 2)+ Math.pow(getZ(), 2));
+	public float length(){
+		return (float)Math.sqrt(Math.pow(getX(), 2) + Math.pow(getY(), 2)+ Math.pow(getZ(), 2));
 	}
 	
 	public void normalize(){
-		double length = length();
+		float length = length();
 		setX(getX() / length);
 		setY(getY() / length);
 		setZ(getZ() / length);
 	}
 	
-	public double dotProduct(Vector other){
+	public float dotProduct(Vector other){
 		return getX()*other.getX() + getY()*other.getY() + getZ()*other.getZ();
 	}
 	
-	public double dotProductWithPoint(Point other){
+	public float dotProductWithPoint(Point other){
 		return getX()*other.getX() + getY()*other.getY() + getZ()*other.getZ();
 	}
 	
 	public Vector crossProduct(Vector other){
-		double newX = getY()*other.getZ() - getZ()*other.getY();
-		double newY = getZ()*other.getX() - getX()*other.getZ();
-		double newZ = getX()*other.getY() - getY()*other.getX();
+		float newX = getY()*other.getZ() - getZ()*other.getY();
+		float newY = getZ()*other.getX() - getX()*other.getZ();
+		float newZ = getX()*other.getY() - getY()*other.getX();
 		return new Vector(newX,newY,newZ);
 	}
 	
 	public Vector add(Vector other){
-		double newX = getX()+other.getX();
-		double newY = getY()+other.getY();
-		double newZ = getZ()+other.getZ();
+		float newX = getX()+other.getX();
+		float newY = getY()+other.getY();
+		float newZ = getZ()+other.getZ();
 		return new Vector(newX,newY,newZ);
 	}
 	
-	public Vector multiply(double scalar){
-		double newX = getX()*scalar;
-		double newY = getY()*scalar;
-		double newZ = getZ()*scalar;
+	public Vector multiply(float scalar){
+		float newX = getX()*scalar;
+		float newY = getY()*scalar;
+		float newZ = getZ()*scalar;
 		return new Vector(newX,newY,newZ);
 	}
